Adding authentication logic to app. Currently hardcoded to a single ID, but can now begin adding better-auth support.
Added UserContext and included it at the top level in my providers wrapper. It was originally created to make development easier, as I didn't want to revert BetterAuth routes to work with localhost, but it seems to be a decent solution for easily storing User data after the SSO/OAuth handshake.
ex: const { isLoggedIn, userId } = useUserContext();

Updated routes to use Prisma.

Updated the [projectId] route to update all associated tasks during updates.

Fixed rollback methods in the ProjectContext.

Updated create, update, and delete operations in ProjectContext so that private/public state variables are updated correctly.

Modified ProjectChooser and ProjectsPage to ensure only a single '+ New Project' button is shown at a time.

Created a Header component and added it to the root layout. It shows the user if logged in and contains a new NavBar component to make navigation easier.

Moved context providers to the root layout, meaning the providers are now available on the landing page.

Updated next.config.ts to use images: { remotePatterns: [...] } instead of the deprecated domains: [...].

Updated definitions.tsx. Optionally null props are now defined as:
prop?: type | null;
instead of:
prop: type | null;
This technically allows the value of undefined to be set for the prop. This is acceptable because our prisma.schema defines optionally null props like so:
prop type?
This means undefined values are set to NULL by default and, by extension, allows NULL to be explicitly set as a valid value.

Added api/projects/recent/route.ts to mirror the same route that exists for Tasks. This route is now used in LiveDataPreview.

Updated api/projects/route.ts. The POST request now includes an added userId property. The GET request now returns:

text
{
  success: true,
  projects: projects,     // Will be removed once full implementation is complete
  public: publicProjects, // Defined by a NULL userId
  private: privateProjects
}

Updated ProjectContext to split fetched projects into publicProjects and privateProjects.

Added a checkbox to ProjectEditForm and ProjectForm to mark a project as public.

Updated ProjectChooser to no longer directly use ProjectContext. Instead, it now accepts a Projects[] array as a parameter.

Added todo.txt to gitignore